+++ Source/ArtilleryRuntime/Public/Systems/ArtilleryDispatch.h
@@ -14,6 +14,7 @@
 #include "FArtilleryTicklitesThread.h"
 #include "KeyCarry.h"
 #include "TransformDispatch.h"
+#include "PhysicsTypes/BarragePlayerAgent.h"
 #include "ArtilleryDispatch.generated.h"
 
 
@@ -74,6 +75,7 @@ class ARTILLERYRUNTIME_API UArtilleryDispatch : public UTickableWorldSubsystem
 	friend class UArtilleryLibrary;
 protected:
 	static inline UArtilleryDispatch* SelfPtr = nullptr;
+
 public:
 	
 	OnArtilleryActivated BindToArtilleryActivated;
@@ -322,124 +324,3 @@ private:
 	FSharedEventRef StartTicklitesApply;
 };
 
-UCLASS(meta=(ScriptName="AbilitySystemLibrary"))
-class ARTILLERYRUNTIME_API UArtilleryLibrary : public UBlueprintFunctionLibrary
-{
-	GENERATED_BODY()
-public:
-	UFUNCTION(BlueprintCallable, meta = (ScriptName = "GetAttribute", DisplayName = "Get Attribute Of", ExpandBoolAsExecs="bFound"), Category="Artillery|Attributes")
-	static float K2_GetAttrib(FSkeletonKey Owner, E_AttribKey Attrib, bool& bFound)
-	{
-		bFound = false;
-		return implK2_GetAttrib(Owner,Attrib, bFound);
-	}
-
-	static float implK2_GetAttrib(FSkeletonKey Owner, E_AttribKey Attrib, bool& bFound)
-	{
-		
-		bFound = false;
-		if(UArtilleryDispatch::SelfPtr)
-		{
-			if(UArtilleryDispatch::SelfPtr->GetAttrib( Owner, Attrib))
-			{
-				bFound = true;
-				return UArtilleryDispatch::SelfPtr->GetAttrib( Owner, Attrib)->GetCurrentValue();
-			}
-		}
-		return NAN;
-	}
-	UFUNCTION(BlueprintCallable, meta = (ScriptName = "GetRelatedKey", DisplayName = "Get Related Key From", ExpandBoolAsExecs="bFound"), Category="Artillery|Keys")
-	static FSkeletonKey K2_GetIdentity(FSkeletonKey Owner, E_IdentityAttrib Attrib, bool& bFound)
-	{
-		
-		bFound = false;
-		return implK2_GetIdentity(Owner, Attrib, bFound);
-	}
-	
-	static FSkeletonKey implK2_GetIdentity(FSkeletonKey Owner, E_IdentityAttrib Attrib, bool& bFound)
-	{
-		
-		bFound = false;
-		if(UArtilleryDispatch::SelfPtr)
-		{
-			auto ident = UArtilleryDispatch::SelfPtr->GetIdent( Owner, Attrib);
-			if(ident)
-			{
-				bFound = true;
-				return ident->CurrentValue;
-			}
-		}
-		return FSkeletonKey();
-	}
-
-	UFUNCTION(BlueprintCallable, meta = (ScriptName = "GetPlayerRelatedKey", DisplayName = "Get Local Player's Related Key", WorldContext = "WorldContextObject", HidePin = "WorldContextObject", ExpandBoolAsExecs="bFound"),  Category="Artillery|Keys")
-	static FSkeletonKey K2_GetPlayerIdentity(UObject* WorldContextObject, E_IdentityAttrib Attrib, bool& bFound)
-	{
-		
-		bFound = false;
-		auto ptr = WorldContextObject->GetWorld()->GetSubsystem<UCanonicalInputStreamECS>();
-		if(ptr)
-		{
-			auto streamkey = ptr->GetStreamForPlayer(PlayerKey::CABLE);
-			auto key = ptr->ActorByStream(streamkey);
-			if(key)
-			{
-				return  implK2_GetIdentity(key, Attrib, bFound);
-			}
-		}
-		bFound = false;
-		return FSkeletonKey();
-	}
-
-	UFUNCTION(BlueprintCallable, meta = (ScriptName = "GetThisActorAttribute", DisplayName = "Get My Actor's Attribute", DefaultToSelf = "Actor", HidePin = "Actor", ExpandBoolAsExecs="bFound"),  Category="Artillery|Attributes")
-	static float K2_GetMyAttrib(AActor *Actor, E_AttribKey Attrib, bool& bFound)
-	{
-	
-		bFound = false;
-		auto ptr = Actor->GetComponentByClass<UKeyCarry>();
-		if(ptr)
-		{
-			if(FSkeletonKey key = ptr->GetObjectKey())
-			{
-				return implK2_GetAttrib(key, Attrib, bFound);
-			}
-		}
-		bFound = false;
-		return NAN;
-	}
-
-	UFUNCTION(BlueprintCallable, meta = (ScriptName = "GetPlayerAttribute", DisplayName = "Get Local Player's Attribute", WorldContext = "WorldContextObject", HidePin = "WorldContextObject", ExpandBoolAsExecs="bFound"),  Category="Artillery|Attributes")
-	static float K2_GetPlayerAttrib(UObject* WorldContextObject, E_AttribKey Attrib, bool& bFound)
-	{
-		bFound = false;
-		auto ptr = WorldContextObject->GetWorld()->GetSubsystem<UCanonicalInputStreamECS>();
-		if(ptr)
-		{
-			auto streamkey = ptr->GetStreamForPlayer(PlayerKey::CABLE);
-			auto key = ptr->ActorByStream(streamkey);
-			if(key)
-			{
-				return  implK2_GetAttrib(key, Attrib, bFound);
-			}
-		}
-		bFound = false;
-		return NAN;
-	}
-	UFUNCTION(BlueprintCallable, meta = (ScriptName = "GetPlayerAttribute", DisplayName = "Get Local Player's Attribute"),  Category="Artillery|Character")
-	static float GetLocalPlayerFRVectors(UObject* WorldContextObject, E_AttribKey Attrib, bool& bFound)
-	{
-		bFound = false;
-		auto ptr = UCanonicalInputStreamECS::SelfPtr;
-		if(ptr)
-		{
-			auto streamkey = ptr->GetStreamForPlayer(PlayerKey::CABLE);
-			auto key = ptr->ActorByStream(streamkey);
-			if(key)
-			{
-				return  implK2_GetAttrib(key, Attrib, bFound);
-			}
-		}
-		bFound = false;
-		return NAN;
-	}
-};
